backend/requirement.txt
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.10.0
attrs==25.3.0
beautifulsoup4==4.14.2
blinker==1.9.0
blis==1.3.0
branca==0.8.1
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.4.26
cffi==2.0.0
charset-normalizer==3.4.2
click==8.1.8
cloudpathlib==0.21.0
colorama==0.4.6
confection==0.1.5
curl_cffi==0.13.0
cvzone==1.6.1
cymem==2.0.11
dj-database-url==3.0.1
Django==5.2.8
django-heroku==0.3.1
dnspython==2.7.0
email_validator==2.2.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
fastapi==0.123.10
filelock==3.20.0
folium==0.19.5
frozendict==2.4.7
fsspec==2025.12.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==23.0.0
h11==0.16.0
httplib2==0.22.0
huggingface-hub==0.36.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
multitasking==0.0.12
murmurhash==1.0.12
narwhals==1.38.1
networkx==3.6
numpy==2.2.5
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
packaging==24.2
pandas==2.2.3
peewee==3.18.3
pillow==11.2.1
platformdirs==4.5.0
plotly==6.4.0
preshed==3.0.9
proto-plus==1.26.1
protobuf==5.29.4
psycopg2==2.9.11
psycopg2-binary==2.9.11
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.4
pydantic_core==2.33.2
pydeck==0.9.1
pygame==2.6.0
Pygments==2.19.1
pyparsing==3.2.3
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
qrcode==8.0
referencing==0.36.2
regex==2025.11.3
requests==2.32.3
rich==14.0.0
rpds-py==0.24.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.16.3
setuptools==80.3.1
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
soupsieve==2.8
spacy==3.8.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.3
srsly==2.5.1
starlette==0.50.0
streamlit==1.45.0
streamlit_folium==0.25.0
sympy==1.14.0
tenacity==9.1.2
thinc==8.3.6
threadpoolctl==3.6.0
tokenizers==0.22.1
toml==0.10.2
torch==2.9.1
tornado==6.4.2
tqdm==4.67.1
transformers==4.57.3
typer==0.15.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.38.0
wasabi==1.1.3
watchdog==6.0.0
weasel==0.4.1
websockets==15.0.1
whitenoise==6.11.0
wrapt==1.17.2
xyzservices==2025.4.0
yfinance==0.2.66
